agents/requirements.txt
# DrawnBuy Agent Team — Python Dependencies
# Install with: pip install -r agents/requirements.txt

# ── Core agent framework ───────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.14.0

# ── Anthropic Claude (LLM provider) ───────────────────────────────────────
anthropic>=0.34.0
langchain-anthropic>=0.2.0

# ── Environment & configuration ────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Data validation ────────────────────────────────────────────────────────
pydantic>=2.0.0

# ── Web research tools (for Affiliate Agent) ──────────────────────────────
# SerperDevTool requires a free API key from serper.dev
# ScrapeWebsiteTool for scraping affiliate network pages
requests>=2.31.0
beautifulsoup4>=4.12.0

# ── Utilities ─────────────────────────────────────────────────────────────
pathlib2>=2.3.7; python_version < '3.4'
